UPDATE: Build scripts. FIX: Barbedo.cpp. No more conversion warnings. UPDATE: Barbedo. Native dll name and location. UPDATE: ExtensionLoader. New extension directory. UPDATE: MusiCObject. Re-enabled logging.

+++ src/Extensions/Classifiers/Barbedo/Barbedo.cpp
@@ -201,7 +201,7 @@ extern "C"
 			int loop_min = 0;
 
             // Create combination control structure
-			tRefVecIndex * refVecsIndex = mc_tRefVecsIndex_alloc(fClassA->nFrames, fClassB->nFrames);
+			tRefVecIndex * refVecsIndex = mc_tRefVecsIndex_alloc((long)fClassA->nFrames, (long)fClassB->nFrames);
 
 			if(refVecsIndex == NULL)
 			{
